requirements.txt
playwright
beautifulsoup4
botasaurus
pymongo
tqdm
google-genai
python-dotenv
scikit-learn
python-pptx
wordcloud